api/requirements.txt
flask==3.0.2
python-dotenv==1.0.0
pyyaml==6.0.1
flask-cors==4.0.0
pytest==8.0.0
pytest-flask==1.3.0 